Title: 합과 평균, Time: 5ms, Memory: 0MB, Status: Wrong Answer - CodeTree

+++ 240119/합과 평균/sum-and-avg.cpp
@@ -1,7 +1,7 @@
 #include <iostream>
 using namespace std;
 int main() {
-    double a, b;
+    int a, b;
     cin >> a >> b;
     cout << a+b << " " << (double)(a+b)/2;
     // 여기에 코드를 작성해주세요.
